Add files via upload
So I am swamped on an issue...I attempted to split the 'incident_number' column into two so that I can use the numbers before the '-' as the incident year, and the numbers after the '-' as the incident number. So far, the code I used produced an error...does anyone know how to overcome this?